[oh-tab-zhwi.4.4] Split llmProfiles host handlers by concern
